third_party/blink/web_tests/external/wpt/css/css-values/calc-size/calc-size-typed-om-expected.txt

This is a testharness.js-based test.
[FAIL] computedStyleMap serialization of max-width: calc-size(max-content, size / 2)
  assert_equals: expected "calc-size(max-content, (0.5 * size))" but got "calc-size(max-content, 0.5 * size)"
Harness: the test ran to completion.
